Rearrange audio controller and ws protocol modules
Move the audio controller to the dialogue runtime, as it tracks
per-conversation state.
Repackage the protocol module into multiple smaller modules,
in preparation for adding the audio protocol types
