+++ lib/dialogue-agent/audio/interface.ts
